fix: refactor and add broadcast
This enables the notifier to recieve broadcast messages on an api endpoint, and it also splits the app up in more manageable chunks.
