+++ notifier/pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.9"
 dependencies = [
     "dateparser>=1.2.0",
     "feedparser>=6.0.11",
+    "flask>=3.1.0",
     "humanize>=4.11.0",
     "loguru>=0.7.2",
     "python-dateutil>=2.9.0.post0",
